--- Brainstorm
@@ -1,2 +1,0 @@
-Brainstorming session to figure out what need be done and allocate jobs:
-
