--- README.md
@@ -1,5 +1,0 @@
-# Vertigo_Project
-Trinity College Dublin software engineering team working on a project relating to vetigo treatment.
-
-
-Google Drive: drive.google.com/drive/u/1/folders/1o9p1f7r8HgvXhnbG2ZVwvN7O-M-Y3oKl
